+++ packages/yourloops/test/integration/use-cases/device-settings-visualisation.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2023, Diabeloop
+ * Copyright (c) 2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,40 +25,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import MedicalFilesApi from '../../../lib/medical-files/medical-files.api'
-import { patient1Id } from '../data/patient.api.data'
+import { checkCopyTextButton, checkDeviceSettingsContent } from '../assert/device-page.assert'
 
-export const mockMedicalFilesAPI = (teamId: string = 'teamId', teamName: string = 'teamName') => {
-  jest.spyOn(MedicalFilesApi, 'getMedicalReports').mockResolvedValue([{
-    id: 'medicalReportId',
-    authorId: 'authorId',
-    creationDate: '2022-01-10T08:34:06.898Z',
-    patientId: patient1Id,
-    teamId,
-    teamName,
-    diagnosis: 'whatever diagnosis',
-    progressionProposal: 'whatever proposal',
-    trainingSubject: 'here is the subject',
-    number: 1,
-    authorFirstName: 'Vishnou',
-    authorLastName: 'Lapaix'
-  },
-  {
-    id: 'medicalReportId2',
-    authorId: 'authorId',
-    creationDate: '2022-01-02T10:30:00.000Z',
-    patientId: patient1Id,
-    teamId,
-    teamName,
-    diagnosis: 'whatever diagnosis 2 ',
-    progressionProposal: 'whatever proposal 2',
-    trainingSubject: 'here is the subject 2',
-    number: 2,
-    authorFirstName: 'Vishnou',
-    authorLastName: 'Lapaix'
-  }])
-}
-
-export const mockMedicalFilesApiEmptyResult = () => {
-  jest.spyOn(MedicalFilesApi, 'getMedicalReports').mockResolvedValue([])
+export const testDeviceSettingsVisualisation = async () => {
+  checkDeviceSettingsContent()
+  await checkCopyTextButton()
 }